// app/src/generated/models/Ben_ltkinitiativetemplatesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ben_ltkinitiativetemplatesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ben_ltkinitiativetemplatesstatecode = keyof typeof Ben_ltkinitiativetemplatesstatecode;
export const Ben_ltkinitiativetemplatesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ben_ltkinitiativetemplatesstatuscode = keyof typeof Ben_ltkinitiativetemplatesstatuscode;

export interface Ben_ltkinitiativetemplatesBase {
  ben_active?: boolean;
  ben_boardid?: string;
  ben_company?: string;
  ben_description?: string;
  ben_fieldsjson?: string;
  ben_ltkinitiativetemplateid: string;
  ben_method?: string;
  ben_metricsjson?: string;
  ben_name?: string;
  ben_order?: number;
  ben_rolesjson?: string;
  ben_singleaction?: boolean;
  ben_stagesjson?: string;
  ben_templateid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ben_ltkinitiativetemplatesstatecode;
  statuscode?: Ben_ltkinitiativetemplatesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ben_ltkinitiativetemplates extends Ben_ltkinitiativetemplatesBase {
  ben_activename?: string;
  ben_singleactionname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}